Fix cache issue
Fixing loadPage caching broke edit mode. This is why.

loadContent appends data to an array that was presumed to be empty on
load. A faulty if statement allowed content to be fetched from cache but
then still trigger a loadContent in edit mode. This would convert
expected associative arrays into an array of associative arrays.

then most functions utilizing content would just not find anything in
the expected keys, cause keys were now numeric and not string based keys
as expected and therefore just returned null. PHP is great is it not?